kasparro-agentic-fb-analyst/requirements.txt
pandas==2.2.2
pyyaml==6.0.2
numpy==1.26.4
tqdm==4.66.4
openai==1.35.3      # or whichever LLM client you actually use
python-dotenv==1.0.1
